+++ tiledesk-helpcenter-client/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This library is on npm: https://www.npmjs.com/package/@tiledesk/helpcenter-query-client
 
+### 0.1.8
+- search() added workspaceId as first parameter. workspaceId removed from constructor.
+
+### 0.1.7
+- allWorkspaces() promisified
+- search() promisified
+
+### 0.1.6
+- Added allWorkspaces()
+
 ### 0.1.5
 - README.md fixes